requirements.txt
pandas==2.1.0
numpy==1.26.4
scikit-learn==1.3.0
matplotlib==3.8.0
seaborn==0.13.2
openai>=1.0.0
mlflow==2.7.0
python-jenkins==1.8.2
jenkins-job-builder
streamlit==1.27.0
pytest
python-dotenv
